# api/swagger.yaml
openapi: 3.0.3
info:
  title: Jobot API
  description: |
    REST API для Telegram бота Jobot - системы поиска работы и размещения вакансий.
    
    ## Возможности
    - Управление пользователями (сотрудники и работодатели)
    - Создание и управление профилями
    - Публикация вакансий
    - Загрузка резюме
    - Система лайков (matching)
    
  version: 2.0.0
  contact:
    name: Jobot Team
    url: https://github.com/sashasych/jobot
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.jobot.com
    description: Production server
...
paths:
...
  /api/users:
    post:
      tags:
        - users
      summary: Создать пользователя
      description: Создание нового пользователя Telegram бота
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{userId}:
    get:
      tags:
        - users
      summary: Получить пользователя
      description: Получение пользователя по ID
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID пользователя
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - users
      summary: Обновить пользователя
      description: Обновление данных пользователя
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request
        '404':
          description: User not found
    
    delete:
      tags:
        - users
      summary: Удалить пользователя
      description: Удаление пользователя и всех связанных данных
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted successfully
        '404':
          description: User not found
  /api/users/{userId}/profile:
    get:
      tags:
        - users
      summary: Получить профиль пользователя
      description: Возвращает основную информацию о пользователе и связанные профили сотрудника и/или работодателя
      operationId: getUserProfile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '404':
          description: User not found
...
  /api/employers/{employerId}/vacancies:
    get:
      tags:
        - employers
      summary: Получить вакансии работодателя
      description: Получение всех вакансий конкретного работодателя
      operationId: getEmployerVacancies
      parameters:
        - name: employerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployerVacancyListResponse'
  /api/resumes:
    post:
      tags:
        - resumes
      summary: Загрузить резюме
      operationId: createResume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeCreateRequest'
      responses:
        '201':
          description: Resume created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeResponse'
  /api/resumes/{resumeId}:
    get:
      tags:
        - resumes
      summary: Получить резюме
      operationId: getResumeById
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeResponse'
    
    put:
      tags:
        - resumes
      summary: Обновить резюме
      operationId: updateResume
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeUpdateRequest'
      responses:
        '200':
          description: Resume updated successfully
    
    delete:
      tags:
        - resumes
      summary: Удалить резюме
      operationId: deleteResume
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resume deleted successfully
  /api/vacancies:
    post:
      tags:
        - vacancies
      summary: Создать вакансию
      operationId: createVacancy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VacancyCreateRequest'
      responses:
        '201':
          description: Vacancy created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyResponse'
    
    get:
      tags:
        - vacancies
      summary: Получить все вакансии
      operationId: getAllVacancies
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyListResponse'
  /api/vacancies/{vacancyId}:
    get:
      tags:
        - vacancies
      summary: Получить вакансию
      operationId: getVacancyById
      parameters:
        - name: vacancyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyResponse'
    
    put:
      tags:
        - vacancies
      summary: Обновить вакансию
      operationId: updateVacancy
      parameters:
        - name: vacancyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VacancyUpdateRequest'
      responses:
        '200':
          description: Vacancy updated successfully
    
    delete:
      tags:
        - vacancies
      summary: Удалить вакансию
      operationId: deleteVacancy
      parameters:
        - name: vacancyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vacancy deleted successfully
...
components:
  schemas:
    UserCreateRequest:
      type: object
      required:
        - tg_chat_id
        - role
      properties:
        tg_user_name:
          type: string
          example: john_doe
          description: Имя пользователя в Telegram
        tg_chat_id:
          type: string
          example: "123456789"
          description: Уникальный ID чата Telegram
        is_active:
          type: boolean
          default: true
          description: Активен ли аккаунт
        is_premium:
          type: boolean
          default: false
          description: Премиум статус
        role:
          type: string
          enum: [employee, employer]
          description: Роль пользователя
    UserUpdateRequest:
      type: object
      properties:
        tg_user_name:
          type: string
        is_active:
          type: boolean
        is_premium:
          type: boolean
    UserResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            tg_user_name:
              type: string
            tg_chat_id:
              type: string
            is_active:
              type: boolean
            is_premium:
              type: boolean
            role:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    UserProfileResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/UserResponse/properties/data'
            employee:
              $ref: '#/components/schemas/EmployeeResponse/properties/data'
            employer:
              $ref: '#/components/schemas/EmployerResponse/properties/data'
        message:
          type: string
...
    ResumeCreateRequest:
      type: object
      required:
        - employee_id
        - tg_file_id
      properties:
        employee_id:
          type: string
          format: uuid
        tg_file_id:
          type: string
          example: BAADAgADZAAD1234567890
    ResumeUpdateRequest:
      type: object
      properties:
        tg_file_id:
          type: string
    ResumeResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            resume_id:
              type: string
              format: uuid
            employee_id:
              type: string
              format: uuid
            tg_file_id:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    VacancyCreateRequest:
      type: object
      required:
        - employer_id
        - tags
        - title
        - description
        - location
        - salary
      properties:
        employer_id:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
          example: ["golang", "kubernetes", "senior"]
        title:
          type: string
          example: Senior Backend Developer
        description:
          type: string
          example: We are looking for an experienced developer...
        location:
          type: string
          example: Москва (можно удалённо)
        salary:
          type: string
          example: 250,000 - 350,000 руб/месяц
    VacancyUpdateRequest:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        salary:
          type: string
    VacancyResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            vacansie_id:
              type: string
              format: uuid
            employer_id:
              type: string
              format: uuid
            tags:
              type: array
              items:
                type: string
            title:
              type: string
            description:
              type: string
            location:
              type: string
            salary:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    VacancyListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            vacansies:
              type: array
              items:
                $ref: '#/components/schemas/VacancyResponse/properties/data'
    EmployerVacancyListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            vacansies:
              type: array
              items:
                $ref: '#/components/schemas/VacancyResponse/properties/data'
            employer_id:
              type: string
              format: uuid
...
    SuccessResponse:
      type: object
      properties:
        data:
          type: object
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: integer
        timestamp:
          type: string
          format: date-time